pyproject.toml
[project]
name = "aws-bedrock-speed-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "click>=8.1.7",
    "langchain-aws>=0.2.10",
    "matplotlib>=3.10.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "E", "F", "Q000"]
ignore = []

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.cython-compiler]
include-path = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

